+++ utils/save_data.py
@@ -4,10 +4,9 @@ import os.path
 
 from PyQt4.QtGui import QFileDialog
 from PyQt4.QtCore import QSettings
-# from PyQt4.QtGui import QFileDialog
 
 
-def save_shape(self):
+def save_shape(self, save_message):
     """Function to get a file."""
     settings = QSettings('sewer_assessor', 'qgisplugin')
 
@@ -16,7 +15,12 @@ def save_shape(self):
     except TypeError:
         init_path = os.path.expanduser("~")
     filename = QFileDialog.getSaveFileName(None,
-                                           'Save shapefile',
+                                           save_message,
                                            init_path,
                                            'ESRI shapefile (*.shp)')
+
+    if filename:
+        settings.setValue('last_used_import_path',
+                          os.path.dirname(filename))
+
     return filename